addon/*: move tabWidth definition to local scope.
This seemingly introduces duplication, but IMO this constant is a style implementation detail that doesn't make much sense to share between different places.
Column-based pretty-printing looks depend on the exact data being shown. If we think that it's the same data being printed in both places, then we should extract & share a bigger chunk of code than just one const.