init: Don't add quotes to arguments
Quotes are integral part of the flux argv[i] argument as it's directly used by
one of the exec variants and not interpreted by a shell beforehand. Don't add
it! This was leading to errors later on in flux:

```
git repo not ready: git clone --mirror: fatal: protocol '\"https' is not supported
```