assets generator: Fix Go package layout, prevent Go module flicker
Before this PR, `make` caused vfsgen to be added to `go.mod`, and a
consecutive `go mod tidy` reverted that change. This was caused by the
generator not being a valid Go package (because `package main` was
present in a directory with `package assets` files, which confused `go
mod`).

This PR moves the generator to a proper package of its own, removes the
"build" tag which is now unnecessary, and updates go.mod accordingly.