increase timeout now that integration tests have doubled in length
from my observations locally and on the CI, it takes about 18 minutes for one of the tests to run.  Rounding up to 20 that gives us 20 minutes per test (one for centos and one for ubuntu presently) so therefore 50 minutes (adding an extra 5 minutes per test for network latency (if any) on image pulls, as well as slow CI boxes).