Remove resource limits from controller manifest
The 30MB memory limit was quite tight to typical usage: such a limit
can cause paging and then get the process OOM-killed.

The cpu limit was more reasonable, but without any particular reason to
limit this process compared to all the others we run, I think it's
simplest to just eliminate the limits.

Also remove the same text from manifests in tests to avoid confusion.